# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Nairobi Flash Deals API — premium same-day deal platform for upscale venues
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues:
    get:
      operationId: listVenues
      tags: [venues]
      summary: List approved venues
      parameters:
        - in: query
          name: category
          schema:
            $ref: "#/components/schemas/VenueCategory"
        - in: query
          name: neighborhood
          schema:
            $ref: "#/components/schemas/VenueNeighborhood"
        - in: query
          name: limit
          schema:
            type: number
        - in: query
          name: offset
          schema:
            type: number
      responses:
        "200":
          description: List of venues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueListResponse"
    post:
      operationId: createVenue
      tags: [venues]
      summary: Submit a venue for approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVenueRequest"
      responses:
        "201":
          description: Venue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "400":
          $ref: "#/components/responses/BadRequest"
  /venues/{id}:
    get:
      operationId: getVenue
      tags: [venues]
      summary: Get a single venue
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Venue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateVenue
      tags: [venues]
      summary: Update venue details
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVenueRequest"
      responses:
        "200":
          description: Updated venue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "404":
          $ref: "#/components/responses/NotFound"
  /venues/{id}/approve:
    post:
      operationId: approveVenue
      tags: [venues]
      summary: Admin — approve a venue
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Venue approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
  /venues/{id}/analytics:
    get:
      operationId: getVenueAnalytics
      tags: [venues]
      summary: Venue performance analytics
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueAnalytics"

  # ── Deals ────────────────────────────────────────────────────
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: number
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
      required: [message]
    Pagination:
      type: object
      properties:
        total:
          type: number
        limit:
          type: number
        offset:
          type: number
      required: [total, limit, offset]
    VenueCategory:
      type: string
      enum: [restaurant, spa, bar, fitness, experience]
    VenueNeighborhood:
      type: string
      enum: [westlands, kilimani, cbd, karen, langata, lavington, kileleshwa, runda, muthaiga, gigiri, upper_hill, other]
    VenueStatus:
      type: string
      enum: [pending_approval, approved, suspended]
    DealStatus:
      type: string
      enum: [draft, live, filling_fast, sold_out, expired, cancelled]
    DealCategory:
      type: string
      enum: [lunch, dinner, brunch, treatment, class, experience, drinks, tasting]
...
    Venue:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        slug:
          type: string
        category:
          $ref: "#/components/schemas/VenueCategory"
        neighborhood:
          $ref: "#/components/schemas/VenueNeighborhood"
        address:
          type: string
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
        description:
          type: string
        coverImage:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        commissionRate:
          type: string
        status:
          $ref: "#/components/schemas/VenueStatus"
        averageRating:
          type: string
          nullable: true
        totalRatings:
          type: number
        totalBookings:
          type: number
        fillRate:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - category
        - neighborhood
        - address
        - description
        - images
        - commissionRate
        - status
        - totalRatings
        - totalBookings
        - tags
        - createdAt
        - updatedAt
    CreateVenueRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        category:
          $ref: "#/components/schemas/VenueCategory"
        neighborhood:
          $ref: "#/components/schemas/VenueNeighborhood"
        address:
          type: string
        description:
          type: string
        whatsappNumber:
          type: string
        contactEmail:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [name, slug, category, neighborhood, address, description]
    UpdateVenueRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        coverImage:
          type: string
        images:
          type: array
          items:
            type: string
        whatsappNumber:
          type: string
        contactEmail:
          type: string
        websiteUrl:
          type: string
        tags:
          type: array
          items:
            type: string
    VenueListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Venue"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]
    VenueAnalytics:
      type: object
      properties:
        venueId:
          type: number
        totalDeals:
          type: number
        totalBookings:
          type: number
        fillRate:
          type: number
        averageRating:
          type: number
        totalRevenue:
          type: number
        topDealCategory:
          type: string
        bestDay:
          type: string
        bestHour:
          type: number
        recentDeals:
          type: array
          items:
            $ref: "#/components/schemas/Deal"
      required:
        - venueId
        - totalDeals
        - totalBookings
        - fillRate
        - averageRating
        - totalRevenue
        - recentDeals
    Deal:
      type: object
      properties:
        id:
          type: number
        venueId:
          type: number
        title:
          type: string
        description:
          type: string
        category:
          $ref: "#/components/schemas/DealCategory"
        discountPercent:
          type: number
        originalPrice:
          type: string
        dealPrice:
          type: string
        totalSlots:
          type: number
        bookedSlots:
          type: number
        availableSlots:
          type: number
        status:
          $ref: "#/components/schemas/DealStatus"
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        imageUrl:
          type: string
          nullable: true
        isStanding:
          type: boolean
        viewCount:
          type: number
        venue:
          $ref: "#/components/schemas/Venue"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - venueId
        - title
        - description
        - category
        - discountPercent
        - originalPrice
        - dealPrice
        - totalSlots
        - bookedSlots
        - availableSlots
        - status
        - startsAt
        - endsAt
        - isStanding
        - viewCount
        - createdAt
        - updatedAt
...