# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Flash Deals API — premium same-day deal platform for upscale venues
 * OpenAPI spec version: 0.2.0
 */
import type { DealCategory } from "./dealCategory";
import type { DealStatus } from "./dealStatus";
import type { Venue } from "./venue";

export interface Deal {
  id: number;
  venueId: number;
  title: string;
  description: string;
  category: DealCategory;
  discountPercent: number;
  originalPrice: string;
  dealPrice: string;
  totalSlots: number;
  bookedSlots: number;
  availableSlots: number;
  status: DealStatus;
  startsAt: Date;
  endsAt: Date;
  imageUrl?: string | null;
  isStanding: boolean;
  viewCount: number;
  venue?: Venue | null;
  createdAt: Date;
  updatedAt: Date;
}