# lib/api-zod/src/generated/types/error.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Flash Deals API — premium same-day deal platform for upscale venues
 * OpenAPI spec version: 0.2.0
 */

export interface Error {
  message: string;
  code?: string;
}