# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Flash Deals API — premium same-day deal platform for upscale venues
 * OpenAPI spec version: 0.2.0
 */
import type { VenueCategory } from "./venueCategory";
import type { VenueNeighborhood } from "./venueNeighborhood";
import type { VenueStatus } from "./venueStatus";

export interface Venue {
  id: number;
  name: string;
  slug: string;
  category: VenueCategory;
  neighborhood: VenueNeighborhood;
  address: string;
  latitude?: string | null;
  longitude?: string | null;
  description: string;
  coverImage?: string | null;
  images: string[];
  commissionRate: string;
  status: VenueStatus;
  averageRating?: string | null;
  totalRatings: number;
  totalBookings: number;
  fillRate?: string | null;
  tags: string[];
  createdAt: Date;
  updatedAt: Date;
}