# lib/api-zod/src/generated/types/venueAnalytics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Flash Deals API — premium same-day deal platform for upscale venues
 * OpenAPI spec version: 0.2.0
 */
import type { Deal } from "./deal";

export interface VenueAnalytics {
  venueId: number;
  totalDeals: number;
  totalBookings: number;
  fillRate: number;
  averageRating: number;
  totalRevenue: number;
  topDealCategory?: string;
  bestDay?: string;
  bestHour?: number;
  recentDeals: Deal[];
}