# lib/api-zod/src/generated/types/venueCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Flash Deals API — premium same-day deal platform for upscale venues
 * OpenAPI spec version: 0.2.0
 */

export type VenueCategory = (typeof VenueCategory)[keyof typeof VenueCategory];

export const VenueCategory = {
  restaurant: "restaurant",
  spa: "spa",
  bar: "bar",
  fitness: "fitness",
  experience: "experience",
} as const;